+++ src/types.ts
@@ -154,6 +154,11 @@ export type Command =
       type: 'focus'
       selector: string
     }
+  | {
+      type: 'upload',
+      selector: string,
+      files: string[]
+    }
 
 export interface Cookie {
   url?: string